tried to make it work
Hi Mel,
I tried resolving the bugs, but some were tied to incorrect usage of the libraries: I could not find correct replacements for bellman::pairing and also got stuck on fixing the Circuit setup.
Judging from the example that is on the bellman crate main page (https://docs.rs/bellman/0.13.1/bellman/index.html), it seems that there is a lot going on in setting up your own Circuit.